Followup
You were solving the issue attached here. In the implementation you did multiple mistakes:

You should have modified the existing SerenaAgent instead of creating a new class
You also created various new modules which you should not have done. The issue is about extending existing modules and not writing a new mcp module, a new module for a tool, or a new module for an agent.
The test uses too much mocking. I don't want any mocking or monkeypatching. There is a valid serena_agent.config and you can create a temporary project.yaml for tests, otherwise you don't need to mock any tools or anything else. Just check that the existing things work as expected - so be more on the integration test side of things. I forbid you to do monkeypatching!
The prompt extensions in the profiles should be just a single string that is used in templating, no need for a list
Same for the mode, just a single string that is used in the system prompt
You adjusted the autogenerated file prompt_factory.py instead of adjusting the template from which it is created
Have a look at the last commit to see what you did and understand what I'm talking about. Remove all unnecessary new modules that you have written there and correct all the mistakes that I have outlined. Improve the tests. Continue working until the mistakes are fully resolved. Use the symbolic tools to perform minimal edits instead of rewriting whole files!